Fix compilation with rust 1.46+.
Also move the wallet library out of rust-src into mobile_wallet. The rust-src is
the core library, and the wallet is just a wrapper for the mobile wallet, so
there is no need for it to be where it was.
