+++ mobile_wallet/src/ios.rs
@@ -1,5 +1,5 @@
 use libc::c_char;
-use wallet::{
+use crate::{
     check_account_address_ext, combine_encrypted_amounts_ext, create_credential_ext,
     create_encrypted_transfer_ext, create_id_request_and_private_data_ext,
     create_pub_to_sec_transfer_ext, create_sec_to_pub_transfer_ext, create_transfer_ext,
