--- rust-src/wallet/Cargo.toml
@@ -1,80 +1,0 @@
-[package]
-name = "wallet"
-version = "0.1.0"
-authors = ["Concordium AG <developers@concordium.com>"]
-edition = "2018"
-license-file = "../../LICENSE-APACHE"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-pairing = "0.15"
-ff = "0.5"
-rand = "=0.7"
-hex = "0.4"
-serde = "1.0"
-serde_json = "1.0"
-chrono = "0.4"
-ed25519-dalek = "=1.0"
-byteorder = "1.3"
-either = "1.6"
-sha2 = "0.9"
-libc = "0.2"
-anyhow = "1.0"
-
-[dependencies.crypto_common]
-path = "../crypto_common"
-version = "0"
-
-[dependencies.curve_arithmetic]
-path = "../curve_arithmetic"
-version = "0"
-
-[dependencies.elgamal]
-path = "../elgamal"
-version = "0"
-
-[dependencies.dodis_yampolskiy_prf]
-path = "../dodis_yampolskiy_prf"
-version = "0"
-
-[dependencies.id]
-path = "../id"
-version = "0"
-
-[dependencies.pedersen_scheme]
-path = "../pedersen_scheme"
-version = "0"
-
-[dependencies.ps_sig]
-path = "../ps_sig"
-version = "0"
-
-[dependencies.eddsa_ed25519]
-path = "../eddsa_ed25519"
-version = "0"
-
-[dependencies.ffi_helpers]
-path = "../ffi_helpers"
-version = "0"
-
-[dependencies.encrypted_transfers]
-path = "../encrypted_transfers"
-version = "0"
-
-[build-dependencies.elgamal]
-path = "../elgamal"
-version = "0"
-
-[build-dependencies.id]
-path = "../id"
-version = "0"
-
-[build-dependencies.crypto_common]
-path = "../crypto_common"
-version = "0"
-
-[lib]
-name = "wallet"
-crate-type = ["rlib", "cdylib", "staticlib"]
-path = "src/lib.rs"
