+++ rustfmt.toml
@@ -6,7 +6,7 @@ fn_single_line = true
 format_strings = true
 format_macro_matchers = true
 format_macro_bodies = true
-imports_granularity="Crate"
+imports_granularity = "Crate"
 normalize_comments = true
 reorder_impl_items = true
 reorder_imports = true
